Update versions
For the Segment Android integration to be able to
use color customization features only an update
on the WootricSDK used needed to be done. I'm also
 adding CHANGELOG.md and LICENSE.md to the rep.

This means that everytime we change something in
the WootricSDK we also need to update the version
here for it to use the latest. There's probably
a better approach for this so I'll look into it.

No tests we're added since all the code conserning
this integration is already tested.